Solving the polyphony issue. This must be addressed with more care.
As frequencies were set in the oscilator module, any new note would trigger a global change in all oscilators, causing all voices to sample the same frequency.
Now, the frequency is set and copyied to the allocated voice. This garantees that a new note (or a change in the oscilator frequency) does not interfere with previous allocated voices, but it may be a problem with a future live frequency fine-tuning feature. This may be avoided if the frequency fine-tune becomes a property of the wave itself, to be added to the final frequency in the sampling phase. Not a bad approach, but I'm not sure if this is really a good modelling solution.